backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
pytest>=8.0.0
